chore(e2e): improve dem-dev run evidence
